Extract data layer related stuff out side of UI logic
