+++ app/src/main/java/android/support/v17/leanback/supportleanbackshowcase/app/rows/PublishChannelFragment.java
@@ -1,25 +1,27 @@
+/*
+ * Copyright (C) 2017 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package android.support.v17.leanback.supportleanbackshowcase.app.rows;
 
 import android.app.Activity;
-import android.content.ActivityNotFoundException;
-import android.content.ComponentName;
-import android.content.ContentUris;
 import android.content.Context;
 import android.content.Intent;
-import android.database.Cursor;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.graphics.drawable.Drawable;
-import android.media.tv.TvContract;
-import android.net.Uri;
-import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.annotation.DrawableRes;
 import android.support.annotation.NonNull;
-import android.support.media.tv.Channel;
-import android.support.media.tv.ChannelLogoUtils;
-import android.support.media.tv.PreviewProgram;
-import android.support.media.tv.TvContractCompat;
 import android.support.v17.leanback.app.GuidedStepFragment;
 import android.support.v17.leanback.supportleanbackshowcase.R;
 import android.support.v17.leanback.widget.GuidanceStylist;
@@ -27,8 +29,6 @@ import android.support.v17.leanback.widget.GuidedAction;
 import android.util.Log;
 import android.widget.Toast;
 
-
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
@@ -38,7 +38,8 @@ import java.util.List;
  * <p>
  * It extends the GuidedStepFragment to share the similar UI as GuidedStepFragment
  */
-public class PublishChannelFragment extends GuidedStepFragment {
+public class PublishChannelFragment extends GuidedStepFragment
+        implements LoadAddedChannels.Listener {
 
     private static final String TAG = "PublishChannelFragment";
     private static final boolean DEBUG = true;
@@ -52,8 +53,6 @@ public class PublishChannelFragment extends GuidedStepFragment {
      * Bitmap which will be put at the front of each checkbox
      */
     private static final int OPTION_DRAWABLE = R.drawable.row_app_banner;
-
-    private List<ChannelPlaylistId> mPublishedChannelContents = new ArrayList<>();
 
     /**
      * Helper function to add non-checked Action to this fragment
@@ -95,7 +94,7 @@ public class PublishChannelFragment extends GuidedStepFragment {
 
     /**
      * Using different theme as attached activity for consistent UI design requirement
-     *
+     * <p>
      * The theme can be customized in themes.xml
      */
     @Override
@@ -121,7 +120,7 @@ public class PublishChannelFragment extends GuidedStepFragment {
                  * Every time when add channel activity is finished, the LoadAddedChannels async task
                  * will be executed to fetch channels' publish status
                  */
-                new LoadAddedChannels();
+                new LoadAddedChannels(getActivity(), this);
                 Toast.makeText(this.getActivity(), "Channel Added!", Toast.LENGTH_SHORT).show();
             } else {
                 Log.e(TAG, "could not add channel");
@@ -147,7 +146,7 @@ public class PublishChannelFragment extends GuidedStepFragment {
          * executed to fetch channels' publish status
          * And create the guide list accordingly
          */
-        new LoadAddedChannels();
+        new LoadAddedChannels(getActivity(), this);
         return new GuidanceStylist.Guidance(title, description, breadcrumb, icon);
     }
 
@@ -194,233 +193,24 @@ public class PublishChannelFragment extends GuidedStepFragment {
          * to make sure the UI thread will not be blocked
          */
         if (action.isChecked()) {
-            new CreateChannelInMainScreen().execute(selectedChannelContents);
+            new ChannelContents.CreateChannelInMainScreen(getActivity())
+                    .execute(selectedChannelContents);
         } else {
             Toast.makeText(this.getActivity(),
                     getResources().getString(R.string.channel_removed_from_home_screen),
                     Toast.LENGTH_SHORT).show();
-            new RemoveChannelInMainScreen().execute(selectedChannelContents);
-        }
-    }
-
-
-    /**
-     * Async Task to add channel to main screen/ remove channel from main screen
-     */
-
-    /* package */ final class CreateChannelInMainScreen extends AsyncTask<ChannelContents, Void, Long> {
-        private static final String SCHEME = "rowsnewapi";
-        private static final String PACKAGE_NAME
-                = "android.support.v17.leanback.supportleanbackshowcase";
-        private static final String VIDEO_PLAY_ACTIVITY = "playvideo";
-        private static final String MAIN_ACTIVITY = "startapp";
-        public static final String CONTENT_ANDROID_MEDIA_TV_PREVIEW_PROGRAM
-                = "content://android.media.tv/preview_program";
-
-        /**
-         * Channel Logo for published program
-         */
-        private void createChannelLogo(Context context, long channelId,
-                                       @DrawableRes int drawableId) {
-            Bitmap bitmap = BitmapFactory.decodeResource(context.getResources(), drawableId);
-            ChannelLogoUtils.storeChannelLogo(context, channelId, bitmap);
-        }
-
-        /**
-         * The input Id is necessary so user can use the program in home screen to go back to this
-         * application
-         */
-        private String createInputId(Context context) {
-            ComponentName cName = new ComponentName(context, DynamicVideoRowsActivity.class.getName());
-            return TvContractCompat.buildInputId(cName);
-        }
-
-        /**
-         * A transaction id will be assigned to the added channel
-         */
-        private long addChannel(Context context, ChannelContents channelContents) {
-            // This ID is associated with current context
-            String channelInputId = createInputId(context);
-            Channel channel = new Channel.Builder()
-                    .setDisplayName(channelContents.getName())
-                    .setDescription(channelContents.getDescription())
-                    .setType(TvContractCompat.Channels.TYPE_PREVIEW)
-                    .setInputId(channelInputId)
-                    .setAppLinkIntentUri(Uri.parse(SCHEME + "://" + PACKAGE_NAME
-                            + "/" + MAIN_ACTIVITY))
-                    .setInternalProviderId(channelContents.getChannelContentsId())
-                    .build();
-
-            /**
-             * The interaction between current app and launcher interface is through ContentProvider
-             * All the published data should be inserted into content provider
-             */
-            Uri channelUri = context.getContentResolver().insert(TvContractCompat.Channels.CONTENT_URI,
-                    channel.toContentValues());
-            if (channelUri == null || channelUri.equals(Uri.EMPTY)) {
-                Log.e(TAG, "Insert channel failed");
-                return 0;
-            }
-            long channelId = ContentUris.parseId(channelUri);
-            channelContents.setChannelPublishedId(channelId);
-
-            createChannelLogo(context, channelId, R.drawable.row_app_icon);
-
-            List<VideoContent> videos = channelContents.getVideos();
-
-            int weight = videos.size();
-            for (int i = 0; i < videos.size(); ++i, --weight) {
-                VideoContent clip = videos.get(i);
-                final String clipId = clip.getVideoId();
-
-                /**
-                 * Enable preview feature for the video added to the home screen
-                 */
-                PreviewProgram program = new PreviewProgram.Builder()
-                        .setChannelId(channelId)
-                        .setTitle(clip.getTitle())
-                        .setDescription(clip.getDescription())
-                        .setPosterArtUri(Uri.parse(clip.getCardImageUrl()))
-                        .setIntentUri(Uri.parse(SCHEME + "://" + PACKAGE_NAME
-                                + "/" + VIDEO_PLAY_ACTIVITY + "/" + clipId))
-                        .setPreviewVideoUri(Uri.parse(clip.getPreviewVideoUrl()))
-                        .setInternalProviderId(clipId)
-                        .setWeight(weight)
-                        .setPosterArtAspectRatio(clip.getAspectRatio())
-                        .setType(TvContractCompat.PreviewPrograms.TYPE_MOVIE)
-                        .build();
-
-                Uri preview_uri =
-                        Uri.parse(CONTENT_ANDROID_MEDIA_TV_PREVIEW_PROGRAM);
-                Uri programUri = context.getContentResolver().insert(preview_uri,
-                        program.toContentValues());
-                if (programUri == null || programUri.equals(Uri.EMPTY)) {
-                    Log.e(TAG, "Insert program failed");
-                } else {
-                    clip.setProgramId(ContentUris.parseId(programUri));
-                }
-            }
-            return channelId;
-        }
-
-        @Override
-        protected Long doInBackground(ChannelContents... params) {
-            return addChannel(PublishChannelFragment.this.getActivity(), params[0]);
-        }
-
-        @Override
-        protected void onPostExecute(Long channelId) {
-            Intent intent = new Intent(TvContract.ACTION_REQUEST_CHANNEL_BROWSABLE);
-            intent.putExtra(TvContractCompat.EXTRA_CHANNEL_ID, channelId);
-            try {
-                startActivityForResult(intent, ADD_CHANNEL_REQUEST);
-            } catch (ActivityNotFoundException e) {
-                Log.e(TAG, "could not start add channel approval UI", e);
-            }
+            new ChannelContents.RemoveChannelInMainScreen(getActivity(), this)
+                    .execute(selectedChannelContents);
         }
     }
 
     /**
-     * Async Task to remove channel from home screen
-     */
-    private final class RemoveChannelInMainScreen extends AsyncTask<ChannelContents, Void, Void> {
-
-        private void deleteChannel(Context context, long channelId) {
-            int rowsDeleted = context.getContentResolver().delete(
-                    TvContractCompat.buildChannelUri(channelId), null, null);
-            if (rowsDeleted < 1) {
-                Log.e(TAG, "Delete channel failed");
-            }
-        }
-
-        @Override
-        protected Void doInBackground(ChannelContents... params) {
-            ChannelContents playlist = params[0];
-            deleteChannel(getActivity(), playlist.getChannelId());
-            new LoadAddedChannels();
-            return null;
-        }
-    }
-
-    /**
-     * ContentProvider project scheme
-     */
-    private static final String[] CHANNELS_MAP_PROJECTION =
-            {TvContractCompat.Channels._ID, TvContractCompat.Channels.COLUMN_INTERNAL_PROVIDER_ID};
-
-    /**
-     * Set query index according to the projection map specific to channel
-     */
-    private static final int CHANNELS_COLUMN_ID_INDEX = 0;
-    private static final int CHANNELS_COLUMN_INTERNAL_PROVIDER_ID_INDEX = 1;
-
-    /**
-     * An internal class to group the information of channel Contents ID (internal ID) and published
-     * ID
-     * The reason to have this class is to use channel contents ID (which is unique and won't be
-     * changed for each channel) to locate the published ID (which will be changed alongwith
-     * the transaction on content provider)
-     */
-    static final class ChannelPlaylistId {
-        String mId;
-        long mChannelId;
-
-        ChannelPlaylistId(String id, long channelId) {
-            mId = id;
-            mChannelId = channelId;
-        }
-
-    }
-
-    /**
-     * Async task to load published channel
-     * So the publish status will be preserved every time when the setting activity is created
-     * <p>
-     * Usage: Just create the object (new LoadAddedChannels()), then the async task will be executed
-     */
-    private class LoadAddedChannels extends AsyncTask<Void, Void, Void> {
-        LoadAddedChannels() {
-            this.execute();
-        }
-
-        @Override
-        protected Void doInBackground(Void... params) {
-            loadChannels();
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(Void result) {
-            onPublishedChannelsLoaded(mPublishedChannelContents);
-        }
-    }
-
-    private void loadChannels() {
-        /**
-         * every time when the async task loadChannels is executed, the published channel list
-         * will be updated accordingly
-         */
-        mPublishedChannelContents.clear();
-        try (Cursor cursor = getActivity().getContentResolver().query(TvContract.Channels.CONTENT_URI,
-                CHANNELS_MAP_PROJECTION, null, null, null)) {
-            if (cursor != null) {
-                while (cursor.moveToNext()) {
-                    if (!cursor.isNull(
-                            CHANNELS_COLUMN_INTERNAL_PROVIDER_ID_INDEX)) {
-                        // Found a row that contains a non-null provider id.
-                        String id = cursor.getString(CHANNELS_COLUMN_INTERNAL_PROVIDER_ID_INDEX);
-                        long channelId = cursor.getLong(CHANNELS_COLUMN_ID_INDEX);
-                        mPublishedChannelContents.add(new ChannelPlaylistId(id, channelId));
-                    }
-                }
-            }
-        }
-    }
-
-    /**
+     * Implement the method in LoadAddedChannels.Listener
+     *
      * when the async task is finished, this function will be executed to assign channel's publish
      * status based on if this channel is published or not
      */
+    @Override
     public void onPublishedChannelsLoaded(List<ChannelPlaylistId>
                                                   publishedChannels) {
         HashMap<String, ChannelPlaylistId> loadedChannels = new HashMap<>();
@@ -443,6 +233,5 @@ public class PublishChannelFragment extends GuidedStepFragment {
         }
 
     }
-
 }
